[review-2026-09-12] фиксы 57 находок ревью (BUG/PERF), AC25 build succeeded
Refs: IDEA.md задача 2026-09-12 №3, Reviews/review_2026-09-12.json
Не сделано: Sync.cpp-5 (PERF escalate), Helpers.cpp-7, Dimensions.cpp-1, BrowserPalette.cpp-1/-2/-3, Roombook.cpp-6/-7/-8, TestFunc.cpp-1/-2/-3

+++ Sources/AddOn/ReNum.cpp
@@ -90,7 +90,7 @@ GSErrCode ReNumSelected (SyncSettings &syncSettings) {
     }
     UInt32 qtywrite = paramToWriteelem.GetSize ();
     GS::UniString subtitle = GS::UniString::Printf ("Writing data to %d elements", qtywrite);
-    short i = 2;
+    Int32 i = 2; // APIIo_SetNextProcessPhaseID ожидает Int32* maxval (см. DevKit-25)
     #ifdef ServerMainVers_2700
     bool showPercent = false;
     Int32 maxval = 2;
